+++ common/src/main/kotlin/com/cobblemon/mod/common/battles/ShowdownInterpreter.kt
@@ -560,49 +560,39 @@ object ShowdownInterpreter {
     }
 
     // |move|p1a: Charizard|Tackle|p2a: Magikarp
-    private fun handleMoveInstruction(battle: PokemonBattle, message: String, remainingLines: MutableList<String>) {
-        battle.dispatch {
-            val editMessaged = message.replace("|move|", "")
-            this.lastMover[battle.battleId] = message
-            val userPNX = editMessaged.split("|")[0].split(":")[0].trim()
-            val (_, userPokemon) = battle.getActorAndActiveSlotFromPNX(userPNX)
-            val moveName = editMessaged.split("|")[1].split("|")[0]
-                .replace(" ", "")
-                .lowercase()
-                .replace("[^A-z0-9]".toRegex(), "")
+    private fun handleMoveInstruction(battle: PokemonBattle, rawMessage: String, remainingLines: MutableList<String>) {
+        battle.dispatchGo {
+            this.lastMover[battle.battleId] = rawMessage
+            val message = BattleMessage(rawMessage)
 
-            val moveTemplate = Moves.getByName(moveName) ?: return@dispatch GO.also { battle.broadcastChatMessage("UNRECOGNIZED MOVE: $moveName".text()) }
-            val hasTarget = editMessaged.split("|").size == 3 && editMessaged.split("|")[2].isNotEmpty()
-            val targetPokemon = if (hasTarget) {
-                val targetPNX = editMessaged.split("|")[2].split(":")[0]
-                battle.getActorAndActiveSlotFromPNX(targetPNX)
-            } else {
-                null
-            }
+            val userPokemon = message.actorAndActivePokemon(0, battle)?.second ?: return@dispatchGo
+            val targetPokemon = message.actorAndActivePokemon(2, battle)?.second
+
+            val effect = message.effectAt(1) ?: return@dispatchGo
+            val move = Moves.getByNameOrDummy(effect.id)
+
             userPokemon.battlePokemon?.effectedPokemon?.let { pokemon ->
                 val progress = UseMoveEvolutionProgress()
                 if (progress.shouldKeep(pokemon)) {
-                    val created = pokemon.evolutionProxy.current().progressFirstOrCreate({ it is UseMoveEvolutionProgress && it.currentProgress().move == moveTemplate }) { progress }
+                    val created = pokemon.evolutionProxy.current().progressFirstOrCreate({ it is UseMoveEvolutionProgress && it.currentProgress().move == move }) { progress }
                     created.updateProgress(UseMoveEvolutionProgress.Progress(created.currentProgress().move, created.currentProgress().amount + 1))
                 }
             }
-            if (targetPokemon != null && targetPokemon.second != userPokemon) {
-                val targetPNX = editMessaged.split("|")[2].split(":")[0]
-                val (_, targetPokemon) = battle.getActorAndActiveSlotFromPNX(targetPNX)
+
+            if (targetPokemon != null && targetPokemon != userPokemon) {
                 battle.broadcastChatMessage(battleLang(
                     key = "used_move_on",
                     userPokemon.battlePokemon?.getName() ?: "ERROR".red(),
-                    moveTemplate.displayName,
+                    move.displayName,
                     targetPokemon.battlePokemon?.getName() ?: "ERROR".red()
                 ))
             } else {
                 battle.broadcastChatMessage(battleLang(
                     key = "used_move",
                     userPokemon.battlePokemon?.getName() ?: "ERROR".red(),
-                    moveTemplate.displayName
+                    move.displayName
                 ))
             }
-            GO
         }
     }